Adjusted Player

+++ app/src/main/java/com/example/devs_hackathon_2023/User/Location.java
@@ -13,7 +13,7 @@ public class Location {
         return latitude;
     }
 
-    public void setLatitude(int latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
@@ -21,7 +21,7 @@ public class Location {
         return longitude;
     }
 
-    public void setLongitude(int longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 }
